조해근 : 업데이트

+++ DirectX/DirectX45/GameEngineContents/ServerWindow.cpp
@@ -68,7 +68,7 @@ void ServerWindow::OnGUI(std::shared_ptr<GameEngineLevel> Level, float _DeltaTim
 			size_t StrLen = strlen(TestSendBuffer.c_str());
 
 			// TestSendBuffer = Arr;
-			NetInst->Send(&TestSendBuffer[0], StrLen);
+			NetInst->Send(&TestSendBuffer[0], static_cast<unsigned int>(StrLen));
 		}
 		ImGui::PopID();
